requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
psycopg2-binary==2.9.9
pymongo==4.7.2
boto3==1.34.100
faker==25.0.0
matplotlib==3.9.0
numpy==1.26.4
